Checkpoint 3 - Done minor changes to UI

+++ client/src/components/DepartmentTableComponent.jsx
@@ -2,6 +2,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { FiEdit } from "react-icons/fi";
 import { RiDeleteBin6Line } from "react-icons/ri";
 import { deleteDepartment, getDepartments } from "../features/department/departmentSlice";
+import { Link } from "react-router-dom";
 
 const DepartmentTableComponent = () => {
     const { departments } = useSelector((state) => state.department);
@@ -25,10 +26,14 @@ const DepartmentTableComponent = () => {
                                     <td>{department.deptId}</td>
                                     <td>{department.name}</td>
                                     <td>
+                                        <Link to={`/admin/editDepartment/${department._id}`}>
                                         <button className="btn-secondary btn-edit" style={{ fontSize: '12px' }} ><FiEdit /></button>
+                                        </Link>
                                         <button className="btn-danger btn-delete" style={{ fontSize: '12px' }} onClick={() => {
-                                            dispatch(deleteDepartment(department._id));
-                                            dispatch(getDepartments());
+                                            if (window.confirm(`Are you sure you want to delete ${department.name} department?`)) {
+                                                dispatch(deleteDepartment(department._id));
+                                                dispatch(getDepartments());
+                                            }
                                         }}><RiDeleteBin6Line /></button>
                                     </td>
                                 </tr>
